requirements.txt
langchain==0.0.254
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.25.0
streamlit_chat==0.1.1
